OIM-39: More code shuffling...it's building now!
