+++ JamaaTech.SMPP.Net.Portable/Smpp.Net.Portable/Smpp.Net.Portable.csproj
@@ -49,6 +49,7 @@
     <Compile Include="ByteBuffer.cs" />
     <Compile Include="DataCoding.cs" />
     <Compile Include="Latin1Encoding.cs" />
+    <Compile Include="MessageState.cs" />
     <Compile Include="SmppEncodingUtil.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="SmppErrorCode.cs" />
